TEMP FIX: inlining parts of leaflet.css as a tile alignment quick fix
The deployed version is not finding
`../bower_components/leaflet/dist/leaflet.css` for some reason, which
is making the map tiles fail to tesselate correctly.  Until we figure
out why, pasting in the most important parts of that file to our own
main stylesheet is a quick fix.  THIS SHOULD NOT STAY IN THE CODE!